Move WeakRef to is own submodule
